+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Path variable `@` to [TypeScript configuration](tsconfig.json)
 - [Type declaration](db.d.ts) for user defined links
 - [Header component](components/Header.tsx)
+- [Custom hooks file](utils/hooks.ts) to utils directory
 
 ### Changed